Enable 'User-Agent' for Email Opened + Custom tracker + id

+++ startGoPhish.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 
-ENV_FILENAME=gophish.env
-export $(sed 's/^#//g' ${ENV_FILENAME} | cut -d= -f1)
-docker-compose up -d
-docker image prune -f
+ENV_FILE=gophish.env
+
+if [[ -f ${ENV_FILE} ]]; then
+	source ${ENV_FILE}
+	export $(sed 's/^#//g' ${ENV_FILE} | cut -d= -f1)
+	docker-compose up -d
+	docker image prune -f
+fi
